+++ Phonebook/src/main/java/application/main/controller/AddressController.java
@@ -1,8 +1,8 @@
 package application.main.controller;
 
-import application.main.model.Entity.Address;
+import application.main.model.entity.Address;
 import application.main.service.Dispatcher;
-import application.main.service.Interfaces.IDispatcher;
+import application.main.service.interfaces.IDispatcher;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
