+++ Phonebook/src/main/java/application/main/controller/PersonController.java
@@ -1,8 +1,8 @@
 package application.main.controller;
 
-import application.main.model.Entity.DTOs.SimplePersonDTO;
-import application.main.model.Entity.Person;
-import application.main.service.Interfaces.IDispatcher;
+import application.main.model.entity.dto.SimplePersonDTO;
+import application.main.model.entity.Person;
+import application.main.service.interfaces.IDispatcher;
 import application.main.service.Dispatcher;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
