+++ Phonebook/src/main/java/application/main/model/database/DatabaseHandlerFactory.java
@@ -1,4 +1,4 @@
-package application.main.model.Database;
+package application.main.model.database;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
