+++ Phonebook/src/main/java/application/main/model/database/dao/AddressDAO.java
@@ -1,8 +1,8 @@
-package application.main.model.Database.DAOs;
+package application.main.model.database.dao;
 
-import application.main.model.Database.DatabaseHandlerFactory;
-import application.main.model.Database.Interfaces.IAddressDAO;
-import application.main.model.Entity.Address;
+import application.main.model.database.DatabaseHandlerFactory;
+import application.main.model.database.interfaces.IAddressDAO;
+import application.main.model.entity.Address;
 import com.microsoft.sqlserver.jdbc.SQLServerDriver;
 
 import java.sql.*;
