+++ Phonebook/src/main/java/application/main/model/entity/Address.java
@@ -1,4 +1,4 @@
-package application.main.model.Entity;
+package application.main.model.entity;
 
 import java.util.ArrayList;
