+++ Phonebook/src/main/java/application/main/service/PersonService.java
@@ -1,10 +1,10 @@
 package application.main.service;
 
-import application.main.model.Database.DAOs.PersonDAO;
-import application.main.model.Database.Interfaces.IPersonDAO;
-import application.main.model.Entity.DTOs.SimplePersonDTO;
-import application.main.model.Entity.Person;
-import application.main.service.Interfaces.IPersonService;
+import application.main.model.database.dao.PersonDAO;
+import application.main.model.database.interfaces.IPersonDAO;
+import application.main.model.entity.dto.SimplePersonDTO;
+import application.main.model.entity.Person;
+import application.main.service.interfaces.IPersonService;
 
 import java.util.List;
 import java.util.NoSuchElementException;
