+++ Phonebook/src/test/java/ContactInfoTest.java
@@ -1,4 +1,4 @@
-import application.main.model.Entity.ContactInfo;
+import application.main.model.entity.ContactInfo;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
